Linking Quests and Dialogue, Scene Transitions
•	Input manager for changing between different input types (mouse, keyboard, touch, etc.) seamlessly
•	Incorporated player position into the save system with a PositionSaver component, and removed the parts of the player script that updated the player position. This saves player position on save and on scene change events, instead of having a lot of update calls to always be checking and updating the player position.
•	Made scene changes usable with the save system (deprecating the scenemanager scriptable object from yesterday in favor of something that will work for both saving AND scene transitions)
o	Did this using PlayerPrefs to save previous scene names whenever a new scene is loaded, so the battle scene can always go back to the previous scene when it’s over instead of having to write in all the scenes inside code as more scenes are added to the project)
o	The previous scene is stored in a playerprefs variable called “PreviousScene”, which is set every time a new scene is about to be loaded using the ScenePortal.cs script
•	Added a scene transition manager (StandardSceneTransitionManager.cs) to help make scene transitions smooth always.
o	Relies currently on a SceneFaderCanvas animator gameobject, which is currently in the hierarchy under the nondestroyableobject Quest Machine. This is the only nondestroyable object in the game, it is helpful for save systems and scene transitions.
•	Created a complicated dialogue tree system and example that can read and respond to different quest states:
o
•	Created a functional UI for the quest journal (we’ll want an artist to take a look at all these UI’s I’ve made and redo the artwork. That will be easy, all UI components I’ve used are sprites that she can take and make her own sprites over them)