Inventory Items and GUI
•	Created many item definitions that can be used as examples and starting points for other items. These item definitions are how items are created in game, and all are contained inside the InventoryDatabase.asset object. This completely eliminates the need for “prefabs” as items, no need to duplicate and create a bunch of objects, scriptable or otherwise, for each new item. This will be very helpful as games like this will have many items, many of which will be similar to each other in many ways (hence the item categories I set up last time, which are stackable meaning one item can have many categories)
o	Here is an example of the wooden sword item definition:
o
•	Made changes to input system to allow input from many different types of devices and to allow easy changing of preferred inputs in the future (and possibly someday by the user if they want to remap buttons)
•	Created a “Pickupable Item” prefab. This item can function both as a chest and also as just an item you find in the environment. Simply check or uncheck “is treasure chest” and assign the Item it gives you with the “change” button.
o
o	At the start of the scene, if you have “is treasure chest” unchecked, it will spawn the item to pick up instead of a chest

o	It saves the state of the items/chests that have been picked up when you save the game or change scenes, (and for chests, replaces them with an “opened chest” object which it also saves so you can’t open the chest and get the item again). For items that aren’t chests, it just remembers that that item has been collected and so the item will no longer be there to pick up.

•	Additionally, the items that have been picked up and are now showing up in a rudimentary inventory I also redesigned (I will add graphic elements to it later, just setting up the coding behind it for now):