Save System, Currency, Shopkeeper, Integrations, New Quest
•	SaveSystem updates!
o	Integrated current save system with the inventory, so now it will save inventory, equipment, player position, items in the world that have been picked up/chests opened, quests, and dialogue between npcs in a small JSON file.
o	Created save “slots”
o	And a UI that can handle slots that’s integrated with the inventory system, so it’s all just one pause menu
o
•	Got inventory linked to dialogue and quest systems through primary lua commands, and also some custom actions inside the quest system. I wrote them down in some separate documentation that we’ll give to the game developers such as David and them, but now it works!
o	On that note, I created a quest that directly follows the explore forest quest (and uses that quest’s completion as a prerequisite for beginning the next conversation in the dialogue system, as an example of integrating the different systems).
o	This quest requires the player to obtain two iron swords, and includes examples of how both the dialogue with Dave and the quests can keep track of and interact with the inventory and currency system
o	Created a three-tiered currency system (bronze, silver, gold each one-hundredth of the next) that functions as part of the inventory system
o	Rewards for quests as inventory items and currency demonstrated, as well as checking the inventory or currency during dialogue.
•	Created a shop system.
•	Created an npc that is a shopkeeper, who sells an iron sword, with a dialogue tree that optionally opens the shop.