Bux fixed!
Basically, (and I still have no idea why this was the case, I couldn't find anybody else online who this has happened to or anything), this project was not picking up on any Editor scripts that were buried beneath other folders. That shouldn't happen but it looks like it did anyway. I had to comb through the project and reorganize every editor script into a single editor folder at the root of the assets folder and rewrite just a few scripts that were affected by that. I tested a build and it worked great.