Fixed error where mouse gets stuck in battle scene
The battle scene loads a lot of menus at the same time. Sometimes those menus set the cursor lock state at the same time, so it was a toss-up as to which canvas would win control of the cursor. I now set the cursor lockstate at the beginning of the first battle state, so it will always be unlocked now.