Bug fixes
Level up menu not working - FIXED: for some reason the save changes button had been deleted from the level up script, so it was giving a null reference exception. I added it back in.
Restart game with negative health - FIXED: the save system always wants to have the last say in setting player stats at the beginning of the scene. Usually this is good, except when starting a new game, where it needs to be the "beginningstats" object instead. So I just disabled the saver in GranGranFirst scene.
Quest journal overlapping words - FIXED: wrote a script to dynamically adjust the button height for the quests depending on string length
Spell Menu says "Simple Menu" - FIXED: somehow the aesthetics of the menu hadn't been saved, so the panels and text were all defaults. I did them again.
"Quit Game" and "Close Menu" differentiation - FIXED: easy fix
Move "found gold" banner to not be in the way of the gold - FIXED: another easy one
Gold popup always says "found 2 gold" - FIXED: added a line of code to cause the alert rather than it happening through events that needed to be set manually