More Debugging
shops not working even if items are in the inventory - FIXED: opsive and pixel crushers not playing nice, had to redo draw order of menus
Character sheet not refreshing with items when they are equipped in battle - FIXED: it was only redrawing in the battle scene when the inventory was opened again, added code to redraw windows upon new item equipping as well
Mouse disappears on battle scene sometimes (just make it show the mouse as soon as it starts maybe?) - FIXED (I think): I set the cursor to visible through code at the start of the battle scene in the same place in the code where the cursor is set to unlocked. Seems like that code worked last time for the problem where the cursor would be locked to the centor of the screen, so hopefully this will work for this issue as well.
Added class string to player and included it in the custom saver for Jared
Gran Gran's house music is wrong - I think this is fixed, though I think I fixed it before I was trying to. I think the problem is that the dialogue was starting before the music object was able to run its script, and that was pausing the music object. Fixing an earlier issue with the health starting from negative I think fixed this one too. Feel free to test it again, but I tried testing it and it worked every time for me.